Add the `environment` field to most Python binary target types (#18144)
This change adds the `environment` field to uncontroversial Python binary target types, and fixes its application to `pex_binary`.

Of note, it does _not_ add the field to `python_distribution` (yet), because we might potentially want a more magical solution for Python distributions, given that they are very frequently cross-platform. That same argument _might_ apply to `pex_binary`, but that change would require a deprecation cycle at this point.

Fixes #18141, and fixes #18149.