`shell_command`/`run_in_sandbox` workdir/output dir adjustments (#18157)
This reverts the default behaviour for `experimental_shell_command` wrt the working directory. These targets now default to using the `build_file_dir()` as their default working directory, which seems to make the most sense as a default after letting the adjustable `workdir` field bed in for a while.

This also adds `root_output_directory`, which helps cater for the majority of `relocated_files` uses that arise from running custom commands. The default behaviour is to now output a digest with the working directory's path _stripped_ from all outputs. To retain the old behaviour, you can set the `root_output_directory` to `/`.

Closes #18143
Closes #18110