Fix the dep inference scripts source root bug. (#18164)
The dep inference script used to be a single file that did no importing (except from the stdlib).
And we treated the script as both a resource, for writing into the dep inference sandbox at Pants
run time, and as a Python source, so we can fmt/lint/typecheck it at Pants build time.

But #17997 split up that script to make it more modular, so that it now imports other script code, 
relative to the scripts dir. Now, for the fmt/lint/typecheck to succeed, those imports have to correspond
to a source root, so we added `src/python/pants/backend/python/dependency_inference/scripts/`
as a source root.

But this means that when packaging the `pantsbuild.pants` wheel, those scripts will be placed relative to
the root package, rather than nested in `pants.backend.python.dependency_inference.scripts`. And we 
loading the script contents as resources relative to that package.

This PR:

- Fixes this by removing that nested source root (which is confusing anyway). So those inter-script
imports must now specify the entire `pants.backend.python.dependency_inference.scripts` prefix.
- Gets rid of the `_pants_dep_parser` script package, which was intended to namespace the scripts under
the nested source root, since we now have `pants.backend.python.dependency_inference.scripts` as
a namespace.
- Ensures that all the intermediate `__init__.py` are present in the scripts sandbox, so the scripts
can import on Python 2.
- Gets rid of the resource/code duality in the BUILD file. Python can load source file contents as
resources, so just depending on the scripts as python_sources is fine.