Remove the `boot_script` from `experimental_shell_command`/`experimental_run_in_sandbox` (#18168)
This replaces the boot script that is run on startup for every `esc`/`eris` invocation, and replaces it with a reusable rule that creates a digest full of scripts that call `{tool_name} $@`. This digest is then provided as an `immutable_input_digest` and made available on the `PATH` of the `experimental_shell_command`.

While this doesn't do a whole lot in and of itself, it shows how in future you might be able to replace `tools` with a more general `runnable_dependencies` field, though such a field would be subject to significant design work.