Use `EnvironmentAwarePackageRequest` in many more contexts. (#18203)
Many locations which requested `BuiltPackage` were not doing so in an environment-aware fashion, which meant that the environments of any relevant `PackageFieldSet`s would not be respected. This meant that this example from the documentation was broken: https://www.pantsbuild.org/v2.15/docs/environments#use-a-docker_environment-to-build-the-inputs-to-a-docker_image

Uses `EnvironmentAwarePackageRequest` in more locations (but not in any `run` goals, since we have not yet determined the semantics we want in that case), and add a test for the example case.