Don't specify interpreter_constraints when building the release pex.
We already enumerate all the `--platform`s we want in the pex.
Adding interpreter_constraints will cause us to also try and
add the current platform * py3{7,8,9} into the pex. If we're building
on macos x86_64 or on linux that would be a no-op, since we already
request those platforms explicitly. But for macos arm64 this fails,
since we only release Pants for py39 on that architecture.