Get rid of the LockfileContent class (#18305)
This is a step towards greater unification of user and tool lockfiles.

LockfileContent existed primarily to carry the default lockfile content
for tools, as read from a resource. This led to a lot of 
`Lockfile | LockfileContent` type signatures and
subsequent type-testing on the receiving end. 

Instead, the Lockfile class now takes an URL instead of a path (with
path-like URLs still supported), and we introduce a 
`resource://package/path` URL to represent lockfiles embedded in 
the Pants binary as a resource.