Enable running third-party Go `main` packages (#17449)
Minimal fix for #17427.

This allows `go_third_party_package` to be treated as a potential main, analyzed and executed. I tried to stick to the shape of the code as much as possible so nothing really interesting to say about it. Simply a few branches based on whether the target has a `GoImportPath` instead of a `GoPackageSourcesField` field.

I'm not happy about the test-case using my repo, but didn't find a way of "mocking" a third-party package. If a better method exists, please let me know. Otherwise I can keep the repo around for the foreseeable future, or we can transfer it to the pants org as a test util.