Adds `system_binary` target, which can be used to reproducibly find a pre-installed binary for use with `adhoc_tool` (#18332)
This adds a `system_binary` target to the `adhoc` backend, which allows
you to specify a pre-installed binary as the `runnable` for
`adhoc_tool`. For POSIX-style tools, there's limited value here, but if
there are installable tools which have vastly different semantics
between versions (e.g. `node`), this allows you to run a
version-checking command, and check against a regex for compatibility.

Closes #18180