Docker environment uses a volume for named caches (#18327)
As reported in #18308, using a bind mount for named caches can result in
the cache contents being owned by root on Linux. Additionally, we
suspect that on macOS, hardlinks within the named caches could cause
issues like
https://github.com/pantsbuild/pants/issues/18162#issuecomment-1432238187.

This change moves to using a volume per Docker image hash to store named
caches. This avoids both of the above issues by avoiding bind mounts for
named caches, and should likely also be faster on macOS (where
virtualization of the bind mount filesystem is a performance
bottleneck).

And as reported in #18306, a similar ownership issue also applies (on
Linux) to the bind mount used for inputs and outputs. To resolve that,
this change also `chmod`s outputs (on Linux) to ensure that the host
machine is able to capture them.

Fixes #18308 and fixes #18306.