Make PythonToolBase.pex_requirements() private.
Almost all clients used to_pex_request() to get a PexRequest
for the tool, but a handful of places called pex_requirements()
directly and then constructed a PexRequest manually using information
already available in the PythonToolBase instance.

This change switches all clients to use to_pex_request() and
makes pex_requirements() "private" by prepending an underscore
to its name.

Narrowing the consumption API will make it easier to reason about
the effects of an upcoming change to allow tools to be installed
via a user lockfile.