Errors in `--loop` wait for changes rather than re-running immediately (#18367)
As reported in #16727, errors occurring in a `--loop` currently retry
indefinitely, because we do not preserve the `Generation` value of a
Node in the case where an error is raised.

This change adjusts the signatures of `Graph` methods to preserve the
`Generation` value in case of errors, and additionally allows "pollers"
to be added to errored (`NotStarted`) `Node`s.

This fix is slightly different from what was [suggested on the
issue](https://github.com/pantsbuild/pants/issues/16727#issuecomment-1259782970):
it does _not_ reintroduce memoization of errors, as that would open a
can of worms around "which" errors are safe to memoize. Instead, only
the method signatures involved in `poll`ing are affected, essentially by
transposing `Result<(Value, Generation), ...>` to `(Result<Value, ..>,
Generation)`.

Fixes #16727.