[internal] Don't choke on union types. (#18362)
Closes #18361 

This addresses the most common case with unions that has two plain
types, such as `SomeType | None`.

The unions themselves are likely not to show up as rule types but in
order to reach any sibling attributes of a type we must allow them to
exist.

This also replaces the use of `inspect.get_annotations()` with
`typing.get_type_hints()`. Don't know how I missed that one last time
around.

As there doesn't seem to be _any_ support with parsing `T1 | T2` style
type hints from Python, I've opted for a minor patch in
`typing.ForwardRef` just to not blow up when accessing the type hints,
for the most common case described above.