Support specifying a named resolve as the superset in a PexRequirements. (#18397)
This named resolve must be a user resolve defined in [python].resolves.

This will allow us to move towards specifying a user resolve as the
source to resolve a tool from. We can't use the existing LoadedLockfile
superset for this, because we don't have access to the PythonSetup
subsystem (which provides the list of resolves) when we construct the
PexRequirements for a tool, so we don't know which lockfile the name
corresponds to.