Delegate lockfile requirement validation to Pex when possible. (#18415)
Our lockfile metadata validation code performs various checks. Among
them, it checks that the requirements we want to satisfy from the lockfile
were among those used to generate it.

We perform this check naively - we expect each exact requirement string
to have been used at generation time. But this is overly strict: If we 
generated with `foo==1.2.3` and now we want to satisfy `foo>=1.0.0,<2`
or just `foo` from that lockfile, our check will fail, even though the
lockfile does satisfy that requirement.

Pex, however, will perform precise validation in the course of the
subsetting operation. So this change skips our naive validation when
we know we're about to ask Pex to subset.

We don't skip this when we're using the entire lockfile, with no subsetting,
since in that case we have no other validation.