Add the ability for Pants to provide Python via a union (with a pyenv impl) (#18352)
Fixes #10447

Adding the ability for an optional union (`PythonProvider`) implementer
to return a `PythonExecutable` (newly outfitted with
`append_only_caches`) so that Python for user code can be supplied
completely by rule-code. Additionally added a `pyenv` union
implementation.

The dirty details are:
- Python is essentially NOT relocatable (hats off to IndyGreg). Python
itself, as well as ecosystem tools, are baked with assumptions about
absolute paths.
- Python installation is not atomic. That means we do our work behind a
lock. We run the bash script on every session, but it should be fast as
it only does the work if the target python doesnt exist.

---------

Co-authored-by: Andreas Stenius <git@astekk.se>