Expand on 2.16 shell/adhoc what's new, and some other sections (#18422)
This expands on the "What's new" added in #18416 by:

- expanding on the Shell & Adhoc section. I did a bunch of the testing
and so have a little bit of a sense of what's relevant/helpful for
migrating.
- clarifying the #18390 deprecation: AIUI it's just the default value of
`[python].interpreter_constraints` that's deprecated, not the whole
option. That is, after 2.17, that option will still exist, but be
required. (This copies a chunk of the deprecation text from that PR.)
- adding an example to the Preamble section.

---------

Co-authored-by: Benjy Weinberger <benjy@toolchain.com>