Audit and update requirements for Python tools. (#18431)
#18418 added support for resolving tools from a user resolve. 

Specifically, it added a `requirements` option for each tool that is
expected to contain very loose requirements that we expect
Pants to support. 

We initialized the default value for that option from the existing
`version` and`extra_requirements` fields. However those were often
overly-constrained, sometimes to a single version.

This PR audits all the tools' defaults and in various cases does one or
both of:

1. Update `default_version` to a newer version so we can regen more
up-to-date lockfiles.
2. Explicitly set `default_requirements` to something loose but likely
compatible.

Of course it's hard to know which currently available tool versions
Pants is compatible with, unless we test against them. And it's impossible
to know which future versions Pants will remain compatible with. 

So we assume that tools are semver-adherent, so that floating up in the
same major version is safe. Updating to newer versions lets us at least start
to test against some newer versions, and is also overdue, since many of
our default tool versions are quite old. The proof that these upgrades
are compatible is in the passing tests.

At some point soon we will `deprecate` `version` and `extra_requirements`
in favor of using user lockfiles for custom tool versions.