Let a user resolve shadow a tool lockfile of the same name (Cherry-pick of #18481) (#18493)
This is necessary for the transition to named resolves - as the tool
name will be an obvious choice for many people for the user resolve.

Also, change the name of the synthetic target used for the lockfile.
Currently it's squatting on the resolve name, meaning you can't name
your python_requirements target for your tool by that name. Now we use a
less obtrusive name, that is unlkely to collide with a user-selected
one.