engine: add an option to clear the destination directory when writing digest to disk (#18534)
This PR adjusts the engine function to have
`workspace.write_digest(digest)` extended with the `clear_destination`
boolean argument to indicate that if there's a file/directory in place,
they should be overwritten.

For instance, writing `outdir/foo.txt` into the `outdir/` directory
containing `bar.txt`, should result in having `outdir/` directory with
only `foo.txt`. Currently, it would preserve the `bar.txt` file which
may not be always desirable. In the goal rule, one can perform side
effects, so one could just delete the directory first with pure Python
command (e.g. `os` or `pathlib` or `shutil`), but it feels it would be
tidier if this could be done straight from the Pants engine to avoid
mixing the paradigms.