Materialize "large" files in a new store location and hardlink them in sandboxes (#18153)
Fixes #18048 by having `store` code now store/load from a different root
if the file size is above some threshold. Also fixes #18231 by verifying
content while downloading.

`materialize_file` now hardlinks (falling back to copying when the store
and workdirs are on separate filesystems) directly from this RO on-disk
file if the file doesn't need to be mutable. This means that
`--no-pantsd` and daemon restarts get the benefits of linking.

---------

Co-authored-by: Stu Hood <stuhood@gmail.com>