Fail fast for attempts to use `test --debug` with a docker environment (Cherry-pick of #18560) (#18567)
As described in #18212, `test --debug` currently silently uses the local
environment to attempt to debug a target, which can cause a lot of
confusion. Other goals like `run` at least render a warning.

Until #17182 is fixed, we should ensure that we fail fast for the case
of `test --debug` instead.

Fixes #18212.