Ensure that a symlinked exported venv exists and is valid. (#18575)
Now, if users delete or corrupt the pex_root named cache, rerunning the
export command will fix things.

Also updates the IT to use the non-deprecated pants cli invocation for
export, i.e., enumerating the resolves you want with `--resolve=` rather
than passing specs.

This test was slow, so I also removed some redundancy - we didn't need
to laboriously test that many tools can export. A sampling was enough. 
Now, with the added test, the test file as a whole runs in about the
same time as previously (~100 seconds).

This will be sped up further once we remove the old-style, deprecated
tool lockfiles.