Set the entire env in __run.sh (#18582)
Previously we set the explicit env vars passed to the Process on top of
the existing env. Now we use `env -i` to force the process to run with
exactly the same env as it did in the engine. 

This will create a much more faithful debugging setup when you poke
around in the sandbox and run `__run.sh`.