expose PANTS_VERSION and make it string-comparable (#18573)
Proposed fix for #18528.

Very "dumb" implementation exposing it as a constant to all BUILD files.
Unfortunately `packaging.version.Version` cannot be compared to strings,
so I'm shimming it with a local Version class that allows this. I could
do this in a more localized manner as this'll *affect* all uses of
PANTS_SEMVER - but it seems like it'll reduce boilerplate in a lot of
places. Happy to change based on opinion!

We could also use an explicit comparator, or expose `Version` as well:

```py3 
if pants_version_after("2.13"):
```
or
```py3
if PANTS_VERSION >= Version("2.13")
```