Only materialize immutable files once per process (#18600)
This PR removes `TempImmutableLargeFile` in favor of having the caller
provide a writer function which writes to the `tokio::fs::File`, and we
wrap the writing in the necessary setup/teardown code.

This has a few benefits:
- We can use a `OnceCell` to ensure the file is only trying to
materialized by one thread
- We can cleanup on error now (whoops that was embarrassing)
- Eliminates a heisenbug I was seeing where we'd fail to hardlink with
"No such file or directory".
- I still don't understand the heisenbug, and I'm still worried we'd see
it cross-process :cry:

All in all, big win