narrow pex_binary files warning (Cherry-pick of #18619) (#18628)
There is a helpful warning when using `pex_binary` and one or more of
the transitive dependencies is a `files` target. However, that warning
is overly broad, as there are situations when it is ok to have `files`
targets.

When a PEX depends on a packaged target, such as `python_distribution`,
then any transitive `files` targets are irrelevant because pex just
installs the wheel with whatever is included in it.