Ensure that sandboxed processes exit before their sandboxes are clean… (#18640)
…ed up (Cherry-pick of #18632)

As @jsirois discovered and described in

https://github.com/pantsbuild/pants/issues/16778#issuecomment-1491120170,
because the `local::CommandRunner` does not `wait` for its child process
to have exited, it might be possible for a SIGKILL to not have taken
effect on the child process before its sandbox has been torn down. And
that could result in the process seeing a partial sandbox (and in the
case of #16778, potentially cause named cache corruption).

This change ensures that we block to call `wait` when spawning local
processes by wrapping them in the `ManagedChild` helper that we were
already using for interactive processes. It then attempts to clarify the
contract of the `CapturedWorkdir` trait (but in order to improve
cherry-pickability does not attempt to adjust it), to make it clear that
child processes should fully exit before the `run_and_capture_workdir`
method has returned.

Fixes #16778 🤞.

---------

Co-authored-by: John Sirois <john.sirois@gmail.com>
(cherry picked from commit 1462bef65afa47416999ac7f7358937e538c16e1)

Co-authored-by: Stu Hood <stuhood@gmail.com>