Error when goals used without relevant backends activated (#18420)
I was confused in a new repo when I ran `./pants tailor` and nothing
happened. This is because the goal is not "activated" since no relevant
backend packages had been enabled yet.

This PR reverts https://github.com/pantsbuild/pants/pull/9142. While
no-oping was probably the right decision between v1->v2 transition, it
no longer is. It results in a confusing and silent issue.

We should install eagerly fail, but with helpful instructions.