javascript: Do not read none mapped package.json files in javascript rules (#18523)
Consider the to-be-parsed `TargetGeneratorRequests` and read the
associated `package.json` from globs _before_
`GenerateNodePackageTargets` is consumed to produce `GeneratedTargets`.
It needs to be done like this to avoid a circular rule dependency that
stems from `WrappedTarget` being used to validate addresses, and
`_TargetParametrizations` being an integral part of this validation (I
think?).

This is an attempt at fixing point 1. of the weirdness observed in
https://github.com/pantsbuild/pants/pull/18326#discussion_r1139210901.

I also took some liberties in refactoring
`resolve_target_parametrizations`, because I struggled to navigate it.
It does mean more `@rule`s in the `graph` module though, I dont know if
that is a metric to be weary of?