Add timeout to remote store calls, and adjust name of cache timeout. (#18695)
#16196 moved the cache read timeout down to the network layer, which
made it much more accurate. But cache lookups also involve a number of
calls to a remote `Store`, which did not have their own timeout.

This change adds an RPC timeout for `Store` accesses to allow for
retries of tar-pitted remote store RPCs, and adjusts the naming of the
`--remote-cache-rpc-timeout-millis` option to make it clear that it
applies to all cache operations (including writes).