refactor `package_python_dist` rule so most of the logic can be reused (Cherry-pick of #18701) (#18703)
In #18639, I have a change to allow pants to do editable installs of
`python_distribution`s for in exported-mutable virtualenvs of user code.

I need all of the dist chroot setup that happens in the
`package_python_dist` rule, so I extracted everything that creates
`DistBuildRequest` into a separate helper function.