Switch many tests to use PythonRuleRunner. (#18712)
This is a step towards getting rid of the IC-setting hack introduced
in #18627.

This is a very mechanical change that replaced uses of RuleRunner
with  PythonRuleRunner. There is also one change to a test timeout.

This is not exhaustive - there are other tests that will need similar
treatment before we can rid of the hack.