Adds a fully-qualified path to a Python interpreter for `pex_binary` `RunRequest`s (Cherry-pick of #18699) (#18700)
Fixes #18683 by adding a rule that extracts the `PexFromTargetsRequest`
from our `BuiltPackage` rule for `pex_binary` targets. This allows us to
construct an `InterpreterConstraints` and hence a `PythonExecutable`
identical to that which was used to request the PEX itself.

Adding the indirection to the `BuiltPackage` rule does not have any
apparent impact on `pants package` requests for `pex_binary` targets.