Use address as stable key when batching field sets in `lint`/`test`.
Closes #18566

Using `str(x)` as the stable batching key ends up using `repr` when `x`
is a field set. This previously raised an exception in some cases - it's
also an unintended behavior change introduced as part of the new
partitioning logic in 2.15.x. Revert back to using `.address` as the
partitioning key for field sets.