Use address as stable key when batching field sets in `lint`/`test` (Cherry-pick of #18725) (#18733)
Closes #18566

Using `str(x)` as the stable batching key ends up using `repr` when `x`
is a field set. After #18719 this no longer raises an exception, but
it's still an unintended behavior change introduced as part of the new
partitioning logic in 2.15.x.

Revert back to using `.address` as the partitioning key for field sets,
and add regression tests.