Deprecate "Secondary Ownership" semantics (#18737)
Currently only 3 targets have "secondary owners" fields:
- `pex_binary`
- `python_awslambda` (this assumes the filename is `.py` Doh!)
- `python_google_cloud_function` (this assumes the filename is `.py`
Doh!)

The secondary ownership ability stems from wanting to use a goal by
referring to the file (which presumably the "primary" owner doesn't
support the goal). E.g. `package`ing a Python file packages the relevant
PEX/Lambda/etc....

However, it is a bit short-sighted:
- You can run both a `pex_binary` (secondary) and a `python_source`
(primary). If multiple PEX binaries exist for a single source:
- Goals like `run` and `paths` error on ambiguity and give unhelpful
messages (saying that the filename was ambiguous, and then listing the
filename as a valid address)
- There are no secondary owners outside of Python.

So marking for deprecation and removal. Users can find the owners of the
files from the specs and pipe that into the relevant goal. (If someone
wants, we can add a new goal/flag to aid with this)