Dedup graph errors. (#18770)
This is an incremental improvement in case of rule graph errors.

We can expect a not fully monomorphized graph in case of errors, and as
such the same nodes (with the same errors) may be present multiple times
with different node ids. By removing duplicated error messages from the
resulting list of errors we can reduce the ratio between number of
distinct errors and the number of error messages significantly.

For a simple error I got it down from 15 to 2 errors while for a more
involved error (the same kind but introduced at a point with far more
dependencies) it went down from 3495 to 120.