Use the resolve's ICs for tools.
The default ICs for tools are deliberately loose, because we
currently use them to generate the tool-specific lockfiles.

But if the user chooses to install a tool from a custom resolve
(which may even be an existing user resolve) then the resolve's
lockfile is likely to have much tighter ICs. As a result, the
user then has to tighten the ICs on the tool, so that it's
compatible. This is superfluous config hoop-jumping, since Pants
knows the ICs to use.

This change makes Pants use the ICs configured for the resolve,
unless explicitly overridden.