Use the resolve's ICs for tools. (#18776)
The default ICs for tools are deliberately loose, because we currently
use them to generate the builtin tool lockfiles.

But if the user chooses to install a tool from a custom resolve
(which may even be an existing user resolve) then the resolve's
lockfile is likely to have much tighter ICs. As a result, the user
has to tighten the ICs on the tool, so that it's compatible.

This is superfluous config hoop-jumping, since Pants knows
the ICs to use.

Also, even if using the builtin lockfile, the user will likely want
to use a tighter IC than the very loose one the lockfile was built 
with, and the repowide default seems like the right choice.

This change makes Pants use the ICs configured for the resolve,
unless explicitly overridden. For the builtin lockfile, it will use
the repo-wide default.

Fixes https://github.com/pantsbuild/pants/issues/15358