Default to using the entire tool lockfile. (#18793)
The initial implementation of installing tools from regular user
lockfiles was designed with the feature of subsetting from a
larger lockfile in mind.

Therefore, we introduced the `requirements` option, which
enumerated the requirements to be installed for the tool,
and defaulted to the requirements in the built-in lockfile
provided with Pants.

However now that users have started experimenting with this
feature, it turns out to be quite laborious - in the more common
case where you create a custom lockfile that is intended to be
used in its entirety, you still have to update `requirements`
(e.g., to tweak versions or add extra requirements) even though
you already enumerated the requirements as inputs to lockfile
generation.

This PR changes the semantics of the `requirements` option. 
It now defaults to an empty list, indicating that the entire lockfile
should be used.  Only if you truly need to subset do you enumerate
`requirements` (which can be versionless, since the lockfile provides
the version).

This has the side-effect of no longer validating that a custom tool
lockfile provides a tool version that Pants believes it can handle. 
But this nannying has also been found by users to be annoying, and
is an example of Pants getting in the way (users were just
bypassing this check by updating versions in `requirements`, which
seems like unnecessary hoop-jumping).

Now if a user-provided tool version doesn't work, e.g., because a
CLI arg is no longer available, or other semantics have changed,
then the user will get an error, or unwanted behavior, from the tool,
and it will be up to them to YOLO that.

This change doesn't require deprecation, since the `requirements`
option has not been released yet.