Rename cache to persistent_cache
Cargo audit seems to confuse this local crate with an [upstream crate](https://crates.io/crates/cache) with the same name that was yanked and no longer exists.
this change fixes it.
Not sure why cargo is confused, if there is a better way to resolve this, I am happy to try it.