Increase timeouts of several tests (#18857)
This bumps up the timeouts of several tests that are regularly causing
spurious failures:

1. `src/python/pants/init/load_backends_integration_test.py`
2. `src/python/pants/base/specs_integration_test.py`
3. `src/python/pants/backend/python/goals/coverage_py_integration_test.py`

The most common of these is 1, especially for cherry-picks to 2.16. For
instance, #18832 required 8 re-attempts, #18839 required 5 and #18850
got to 13.

Given these seem to affect 2.16 cherry-picks so much, I'm thinking it
makes sense to cherry pick this back to that branch too, along with the
increase in #18847, and part of #18627 (the increase to
`src/python/pants/engine/streaming_workunit_handler_integration_test.py`).