Add concrete `Context` type to the `graph` crate (#18835)
The `graph` crate currently has a `trait Context`, but no concrete
`Context`. As the use-cases for the `Context` have expanded, many
implementation details of the `Graph` have ended up as public in the
trait, even though they would be better off as private.

This change extracts a concrete `struct graph::Context` which holds an
opaque user context. `graph::Context` has `impl Deref` to the user
context type, which minimizes edits to callsites. There should be no
changes in performance or behavior.

(The primary motivation for this change is that #18855 needs to record
additional information on the `Context`, which felt like the straw that
broke the camel's back and justified splitting out a concrete
`Context`.)